Switch stale reconciler to native controller-runtime controller
