+++ pkg/controllermanager/controller/project/add.go
@@ -23,6 +23,7 @@ import (
 	"github.com/gardener/gardener/pkg/controllermanager/apis/config"
 	"github.com/gardener/gardener/pkg/controllermanager/controller/project/activity"
 	"github.com/gardener/gardener/pkg/controllermanager/controller/project/project"
+	"github.com/gardener/gardener/pkg/controllermanager/controller/project/stale"
 )
 
 // AddToManager adds all Project controllers to the given manager.
@@ -38,6 +39,13 @@ func AddToManager(mgr manager.Manager, cfg config.ControllerManagerConfiguration
 		Config: *cfg.Controllers.Project,
 	}).AddToManager(mgr); err != nil {
 		return fmt.Errorf("failed adding main reconciler: %w", err)
+	}
+
+	if err := (&stale.Reconciler{
+		Config: *cfg.Controllers.Project,
+		Clock:  clock.RealClock{},
+	}).AddToManager(mgr); err != nil {
+		return fmt.Errorf("failed adding stale reconciler: %w", err)
 	}
 
 	return nil
